allows CLAY_LOG_PRETTY to be overridden via arguments
The clay-log documentation states that `CLAY_LOG_PRETTY` is a fallback
value for cases where the `pretty` argument is not explicitly provided
to the initialized logger. This commit addresses a bug where an
explicit `false` argument would be overridden by `CLAY_LOG_PRETTY`.

This commit also addresses the fact that it's not possible to explitly
set `CLAY_LOG_PRETTY` to a `false` value (the string 'false' is
truth-y).