bugfix: prevent plug-ins from loading client-side
When `clay-log` is compiled via `claycli` for use in `kiln` the value
of `CLAY_LOG_PLUGINS` makes it into `client-env.json`. As a result, we
get a ton of errors in the browser console related to plug-in loading.

This commit adds a quick check that will skip plug-in code on the
client.

The ideal solution would be to not include `CLAY_LOG_PLUGINS` at all,
but that seems like a tougher problem to solve with the build tools in `claycli`.