bugfix: prevent plug-ins from loading client-side (#19)
When `clay-log` is compiled via `claycli` for use in `kiln` the value
of `CLAY_LOG_PLUGINS` makes it into `client-env.json`. As a result, we
get a ton of errors in the browser console related to plug-in loading.

This commit adds a quick check that will skip plug-in code on the
client.

The ideal solution would be to not include `CLAY_LOG_PLUGINS` at all,
but that seems like a tougher problem to solve with the build tools in `claycli`.

If we want to support client-side plug-ins in the future, we need to think about using an alternate config value, eg. `CLAY_LOG_CLIENT_PLUGINS` (available at compile-time) along with an import model that is compatible with the `claycli` build tools.