fixes bug which dropped non-enumerable fields from 'data' argument
This commit removes the use of `_.cloneDeep` on additional context
passed to a `log` call (third positional argument).

Cloning `data` introduced a breaking change for individuals passing
objects with non-enumerable properties as the `data` argument.

**Note:**

This reversion re-introduces the risk of mutation when paired with
plug-ins. It also restores the built-in mutation that will append a
`._label` attribute to any `data` argument.

The README was updated to reflect those risks and explain how
they can be managed through careful usage.